new: Preparations for Upload streaming
